# *: propagate pointer to InternalIteratorStats