--- a/error_iter.go
+++ b/error_iter.go
@@ -14,7 +14,7 @@ type errorIter struct {
 }
 
 // errorIter implements the base.InternalIterator interface.
-var _ internalIteratorWithStats = (*errorIter)(nil)
+var _ internalIterator = (*errorIter)(nil)
 
 func newErrorIter(err error) *errorIter {
 	return &errorIter{err: err}
@@ -63,8 +63,6 @@ func (c *errorIter) String() string {
 }
 
 func (c *errorIter) SetBounds(lower, upper []byte) {}
-func (c *errorIter) Stats() InternalIteratorStats  { return InternalIteratorStats{} }
-func (c *errorIter) ResetStats()                   {}
 
 type errorKeyspanIter struct {
 	err error