--- a/external_iterator.go
+++ b/external_iterator.go
@@ -207,7 +207,7 @@ func finishInitializingExternal(it *Iterator) {
 					continue
 				}
 				mlevels = append(mlevels, mergingIterLevel{
-					iter:         base.WrapIterWithStats(pointIter),
+					iter:         pointIter,
 					rangeDelIter: rangeDelIter,
 				})
 			}
@@ -218,12 +218,12 @@ func finishInitializingExternal(it *Iterator) {
 				}
 				sli.init(it.opts)
 				mlevels = append(mlevels, mergingIterLevel{
-					iter:         base.WrapIterWithStats(sli),
+					iter:         sli,
 					rangeDelIter: nil,
 				})
 			}
 		}
-		it.alloc.merging.init(&it.opts, it.comparer.Compare, it.comparer.Split, mlevels...)
+		it.alloc.merging.init(&it.opts, &it.stats.InternalStats, it.comparer.Compare, it.comparer.Split, mlevels...)
 		it.alloc.merging.snapshot = base.InternalKeySeqNumMax
 		it.alloc.merging.elideRangeTombstones = true
 		it.pointIter = &it.alloc.merging