--- a/internal.go
+++ b/internal.go
@@ -33,8 +33,6 @@ type InternalKey = base.InternalKey
 
 type internalIterator = base.InternalIterator
 
-type internalIteratorWithStats = base.InternalIteratorWithStats
-
 // ErrCorruption is a marker to indicate that data in a file (WAL, MANIFEST,
 // sstable) isn't in the expected format.
 var ErrCorruption = base.ErrCorruption