--- a/internal/keyspan/interleaving_iter.go
+++ b/internal/keyspan/interleaving_iter.go
@@ -94,7 +94,7 @@ type SpanMask interface {
 type InterleavingIter struct {
 	cmp         base.Compare
 	comparer    *base.Comparer
-	pointIter   base.InternalIteratorWithStats
+	pointIter   base.InternalIterator
 	keyspanIter FragmentIterator
 	mask        SpanMask
 
@@ -178,7 +178,7 @@ var _ base.InternalIterator = &InterleavingIter{}
 // propagate the bounds down the iterator stack.
 func (i *InterleavingIter) Init(
 	comparer *base.Comparer,
-	pointIter base.InternalIteratorWithStats,
+	pointIter base.InternalIterator,
 	keyspanIter FragmentIterator,
 	mask SpanMask,
 	lowerBound, upperBound []byte,
@@ -1081,18 +1081,6 @@ func (i *InterleavingIter) Close() error {
 // String implements (base.InternalIterator).String.
 func (i *InterleavingIter) String() string {
 	return fmt.Sprintf("keyspan-interleaving(%q)", i.pointIter.String())
-}
-
-var _ base.InternalIteratorWithStats = &InterleavingIter{}
-
-// Stats implements InternalIteratorWithStats.
-func (i *InterleavingIter) Stats() base.InternalIteratorStats {
-	return i.pointIter.Stats()
-}
-
-// ResetStats implements InternalIteratorWithStats.
-func (i *InterleavingIter) ResetStats() {
-	i.pointIter.ResetStats()
 }
 
 func firstError(err0, err1 error) error {